fix: skip text interrupts on expressive agents to avoid races (#376)
* fix: skip text interrupts on expressive agents to avoid races

Text-typed interrupts on expressive (V2) streams caused race conditions
when the user kept typing while the agent was talking — each keystroke
that triggered chat send would also fire an extra interrupt on top of
prior unsettled ones.

Refactor: move per-transport interrupt logic into each StreamingManager
(livekit-manager skips type==='text'; webrtc-manager preserves the
existing validation + payload from sendInterrupt/validateInterrupt).
agent-manager just forwards the call. Removes services/interrupt module.


* lint

---------
