+++ src/services/agent-manager/index.ts
@@ -7,7 +7,6 @@ import {
     ChatResponse,
     ClientToolHandler,
     ConnectionState,
-    CreateSessionV2Options,
     CreateStreamOptions,
     Interrupt,
     Message,
@@ -30,7 +29,6 @@ import { initializeAnalytics } from '../analytics/mixpanel';
 import { interruptTimestampTracker, latencyTimestampTracker } from '../analytics/timestamp-tracker';
 import { createChat, getRequestHeaders } from '../chat';
 import { getInitialMessages } from '../chat/intial-messages';
-import { sendInterrupt, sendInterruptV2, validateInterrupt } from '../interrupt';
 import { SocketManager, createSocketManager } from '../socket-manager';
 import { createMessageEventQueue } from '../socket-manager/message-queue';
 import { StreamingManager } from '../streaming-manager';
@@ -122,12 +120,7 @@ export async function createAgentManager(agent: string, options: AgentManagerOpt
         lastMessage.interrupted = true;
         options.callbacks.onNewMessage?.([...items.messages], 'answer');
 
-        if (isStreamsV2) {
-            sendInterruptV2(items.streamingManager! as StreamingManager<CreateSessionV2Options>);
-        } else {
-            validateInterrupt(items.streamingManager, items.streamingManager?.streamType, videoId);
-            sendInterrupt(items.streamingManager!, videoId!);
-        }
+        items.streamingManager.interrupt(type);
     };
 
     const clientToolHandlers = new Map<string, ClientToolHandler>();
