+++ src/services/streaming-manager/livekit-manager.ts
@@ -4,6 +4,7 @@ import {
     ConnectivityState,
     CreateSessionV2Options,
     CreateStreamOptions,
+    Interrupt,
     Message,
     PayloadType,
     StreamEvents,
@@ -726,6 +727,14 @@ export async function createLiveKitStreamingManager<T extends CreateSessionV2Opt
         unpublishMicrophoneStream,
         publishCameraStream,
         unpublishCameraStream,
+
+        interrupt(type: Interrupt['type']) {
+            // Skip text interrupts for V2/expressive: the orchestrator does not
+            // cancel the in-flight LLM token stream, and an extra interrupt while
+            // a previous one is still settling causes races.
+            if (type === 'text') return;
+            sendDataChannelMessage(JSON.stringify({ topic: DataChannelTopic.Interrupt }));
+        },
 
         registerRpcMethod(method: string, handler: (data: any) => Promise<string>) {
             room?.registerRpcMethod(method, handler);
